requirements.txt
PyQt5==5.15.11
pandas==2.2.3
pyyaml==6.0.2 